Revert "Merge branch 'master' of https://github.com/MisaZhu/KoalaJS"
This reverts commit 6b9922304207afecdf2135baf3f28d14d64dadc5, reversing
changes made to da68e8eddb3ca4c048cbae3de6ef7bceabab476a.

+++ jsvm/Bytecode.cpp
@@ -31,7 +31,7 @@ void Bytecode::add(PC ins) {
 }
 	
 string Bytecode::getStr(int i) {
-	if(i<0 || i == 0xFFFF ||  i>=(int)strTable.size())
+	if(i<0 || i == 0xFFFF ||  i>=strTable.size())
 		return "";
 	return strTable[i];
 }
